minor cleanup for release
