+++ ROADMAP.md
@@ -2,6 +2,7 @@
 
 ### MODEL SERVER / TOOLING
 
+- Improve the kronk model server logging for a chat/completion
 - Solidfy the auth system
   - CLI tooling to create tokens
   - Provide Auth at the endpoint level (completion/embeddings)
